Speed up process by verifying artist album list cache once per script run... (less queries...)

+++ lidarr-automated-downloader.bash
@@ -63,8 +63,7 @@ paths () {
 	
 	if [ ! -d "$LidarrImportLocation" ]; then
 		mkdir -p "$LidarrImportLocation"
-	fi
-	
+	fi	
 }
 
 CleanDLPath () {
@@ -81,6 +80,16 @@ CleanImportPath () {
 	find "${LidarrImportLocation}" -type d -not -newer "${LidarrImportLocation}/cleanup" -exec rm -rf "{}" \; > /dev/null 2>&1
 	rm "${LidarrImportLocation}/cleanup"
 }
+
+CleanCacheCheck () {
+	if [ -d "cache" ]; then
+		if find "cache" -type f -iname "*-checked" | read; then
+			echo "Clearing cache checked verification files"
+			find "cache" -type f -iname "*-checked" -delete
+		fi
+	fi
+}
+
 
 FileAccessPermissions () {
 	echo "Setting file permissions (${FilePermissions})"
@@ -277,15 +286,20 @@ DeezerMatching () {
 
 DownloadList () {
 	# Check cache deezer artistid album list for matching discography album count, if different, delete
-	if [ -f "cache/$DeezerArtistID-albumlist.json" ]; then
-		cachealbumlist="$(cat "cache/$DeezerArtistID-albumlist.json" | jq '.[].id' | wc -l)"
-		if [ "${newalbumlist}" -ne "${cachealbumlist}" ]; then
-			echo "Existing Cached Deezer Artist Album list is out of date, updating..."
-			rm "cache/$DeezerArtistID-albumlist.json"
-			sleep 0.1
-		else
-			echo "Exisiting Cached Deezer Artist (ID: ${DeezerArtistID}) Album List is current..."
+	if [ ! -f "cache/$DeezerArtistID-checked" ]; then
+		if [ -f "cache/$DeezerArtistID-albumlist.json" ]; then
+			cachealbumlist="$(cat "cache/$DeezerArtistID-albumlist.json" | jq '.[].id' | wc -l)"
+			if [ "${newalbumlist}" -ne "${cachealbumlist}" ]; then
+				echo "Existing Cached Deezer Artist Album list is out of date, updating..."
+				rm "cache/$DeezerArtistID-albumlist.json"
+				sleep 0.1
+			else
+				echo "Exisiting Cached Deezer Artist (ID: ${DeezerArtistID}) Album List is current..."
+				touch "cache/$DeezerArtistID-checked"
+			fi
 		fi
+	else
+		echo "Exisiting Cached Deezer Artist (ID: ${DeezerArtistID}) Album List is current..."
 	fi
 	
 	# Cahche deezer artistid album list and save to file for re-use...
@@ -341,10 +355,17 @@ GetDeezerArtistAlbumList () {
 
 	DeezerArtistID=$(echo "${deezeraritstid}" | grep -o '[[:digit:]]*')
 	echo "Deezer Artist ID: $DeezerArtistID"
-	DeezerArtistAlbumList=$(curl -s "https://api.deezer.com/artist/${DeezerArtistID}/albums&limit=1000")
-	newalbumlist="$(echo "${DeezerArtistAlbumList}" | jq ".data | .[] | .id" | wc -l)"
-	if [ -z "$DeezerArtistAlbumList" ] || [ -z "${newalbumlist}" ]; then
-		echo "ERROR: Unable to retrieve albums from Deezer"
+	if [ ! -f "cache/$DeezerArtistID-checked" ]; then
+		DeezerArtistAlbumList=$(curl -s "https://api.deezer.com/artist/${DeezerArtistID}/albums&limit=1000")
+		newalbumlist="$(echo "${DeezerArtistAlbumList}" | jq ".data | .[] | .id" | wc -l)"
+		if [ -z "$DeezerArtistAlbumList" ] || [ -z "${newalbumlist}" ]; then
+			echo "ERROR: Unable to retrieve albums from Deezer"
+		else
+			DownloadList
+		
+			DeezerMatching
+		fi
+			
 	else
 	
 		DownloadList
@@ -879,9 +900,13 @@ CleanDLPath
 
 CleanImportPath
 
+CleanCacheCheck
+
 ProcessLidarrAlbums
 
 CleanDLPath
+
+CleanCacheCheck
 
 #####################################################################################################
 #                                              Script End                                           #
